Fix(CI): rector

+++ inc/replace.class.php
@@ -28,7 +28,7 @@
  * -------------------------------------------------------------------------
  */
 
-use Glpi\Api\Deprecated\Computer_Item;
+
 use Glpi\Asset\Asset_PeripheralAsset;
 
 use function Safe\fclose;
@@ -66,7 +66,9 @@ use function Safe\fwrite;
 class PluginUninstallReplace extends CommonDBTM
 {
     public const METHOD_PURGE              = 1;
+
     public const METHOD_DELETE_AND_COMMENT = 2;
+
     public const METHOD_KEEP_AND_COMMENT   = 3;
 
     public static $rightname = "uninstall:profile";
@@ -152,6 +154,7 @@ class PluginUninstallReplace extends CommonDBTM
                         $itempdf = new $PLUGIN_HOOKS['plugin_pdf'][$type]($olditem);
                         $out = $itempdf->generatePDF([$olditem_id], $tab, 1, false);
                     }
+
                     $name_out .= ".pdf";
 
                 } else {
@@ -165,8 +168,9 @@ class PluginUninstallReplace extends CommonDBTM
                     foreach ($datas as $k => $v) {
                         $out .= $k . ";";
                     }
+
                     $out .= "\r\n";
-                    foreach ($datas as $k => $v) {
+                    foreach ($datas as $v) {
                         $out .= $v . ";";
                     }
 
@@ -211,40 +215,34 @@ class PluginUninstallReplace extends CommonDBTM
             }
 
             // General Information - NAME
-            if ($model->fields["replace_name"]) {
-                if ($overwrite || empty($newitem->fields['name'])) {
-                    $newitem->update(
-                        ['id'  => $newitem_id,
-                            'name' => $olditem->getField('name'),
-                        ],
-                        false,
-                    );
-                }
+            if ($model->fields["replace_name"] && ($overwrite || empty($newitem->fields['name']))) {
+                $newitem->update(
+                    ['id'  => $newitem_id,
+                        'name' => $olditem->getField('name'),
+                    ],
+                    false,
+                );
             }
 
             $data['id'] = $newitem->getID();
             // General Informations - SERIAL
-            if ($model->fields["replace_serial"]) {
-                if ($overwrite || empty($newitem->fields['serial'])) {
-                    $newitem->update(
-                        ['id'     => $newitem_id,
-                            'serial' => $olditem->getField('serial'),
-                        ],
-                        false,
-                    );
-                }
+            if ($model->fields["replace_serial"] && ($overwrite || empty($newitem->fields['serial']))) {
+                $newitem->update(
+                    ['id'     => $newitem_id,
+                        'serial' => $olditem->getField('serial'),
+                    ],
+                    false,
+                );
             }
 
             // General Informations - OTHERSERIAL
-            if ($model->fields["replace_otherserial"]) {
-                if ($overwrite || empty($newitem->fields['otherserial'])) {
-                    $newitem->update(
-                        ['id'          => $newitem_id,
-                            'otherserial' => $olditem->getField('otherserial'),
-                        ],
-                        false,
-                    );
-                }
+            if ($model->fields["replace_otherserial"] && ($overwrite || empty($newitem->fields['otherserial']))) {
+                $newitem->update(
+                    ['id'          => $newitem_id,
+                        'otherserial' => $olditem->getField('otherserial'),
+                    ],
+                    false,
+                );
             }
 
             // Documents
@@ -302,28 +300,24 @@ class PluginUninstallReplace extends CommonDBTM
                 && in_array($type, $CFG_GLPI["infocom_types"])
             ) {
                 $infocom = new Infocom();
-                if ($overwrite) {
-                    // Delete current Infocoms of new item
-                    if ($infocom->getFromDBforDevice($type, $newitem_id)) {
-                        //Do not log infocom deletion in the new item's history
-                        $infocom->dohistory = false;
-                        $infocom->deleteFromDB(true);
-                    }
+                // Delete current Infocoms of new item
+                if ($overwrite && $infocom->getFromDBforDevice($type, $newitem_id)) {
+                    //Do not log infocom deletion in the new item's history
+                    $infocom->dohistory = false;
+                    $infocom->deleteFromDB(true);
                 }
 
                 // Checks that the itemtype/items_id key doesn't already exist to avoid duplication
-                if (!$infocom->getFromDBforDevice($type, $newitem_id)) {
-                    // Update current Infocoms of old item
-                    if ($infocom->getFromDBforDevice($type, $olditem_id)) {
-                        $infocom->update(
-                            [
-                                'id'       => $infocom->getID(),
-                                'itemtype' => $type,
-                                'items_id' => $newitem_id,
-                            ],
-                            false,
-                        );
-                    }
+                // Update current Infocoms of old item
+                if (!$infocom->getFromDBforDevice($type, $newitem_id) && $infocom->getFromDBforDevice($type, $olditem_id)) {
+                    $infocom->update(
+                        [
+                            'id'       => $infocom->getID(),
+                            'itemtype' => $type,
+                            'items_id' => $newitem_id,
+                        ],
+                        false,
+                    );
                 }
             }
 
@@ -402,20 +396,14 @@ class PluginUninstallReplace extends CommonDBTM
 
             // Group
             if (
-                $model->fields["replace_groups"]
-                && in_array($type, $CFG_GLPI["linkgroup_types"])
+                $model->fields["replace_groups"] && in_array($type, $CFG_GLPI["linkgroup_types"]) && ($newitem->isField('groups_id') && ($overwrite || empty($newitem->fields['groups_id'])))
             ) {
-                if (
-                    $newitem->isField('groups_id')
-                    && ($overwrite || empty($newitem->fields['groups_id']))
-                ) {
-                    $newitem->update(
-                        ['id'        => $newitem_id,
-                            'groups_id' => $olditem->getField('groups_id'),
-                        ],
-                        false,
-                    );
-                }
+                $newitem->update(
+                    ['id'        => $newitem_id,
+                        'groups_id' => $olditem->getField('groups_id'),
+                    ],
+                    false,
+                );
             }
 
             // Tickets
@@ -455,11 +443,11 @@ class PluginUninstallReplace extends CommonDBTM
             // Directs connections
             if (
                 $model->fields["replace_direct_connections"]
-                && (in_array($type, ['Computer']))
+                && ($type === 'Computer')
                 && $newitem_id
             ) { #do not update computer_item if no computer
                 $comp_item = new Asset_PeripheralAsset();
-                if (is_a($olditem, Computer::class, true)) {
+                if ($olditem instanceof Computer) {
                     foreach (self::getAssociatedItems($olditem) as $itemtype => $connections) {
                         foreach ($connections as $connection) {
                             $comp_item->update(
@@ -499,27 +487,22 @@ class PluginUninstallReplace extends CommonDBTM
                 if ($model->fields["remove_from_ocs"] == 1) {
                     PluginUninstallUninstall::deleteComputerInOCSByGlpiID($olditem_id);
                 }
+
                 //Delete link in glpi_ocs_link
                 if ($model->fields["delete_ocs_link"] || $model->fields["remove_from_ocs"]) {
                     PluginUninstallUninstall::deleteOcsLink($olditem_id);
                 }
             }
 
-            if ($plug->isActivated('fusioninventory')) {
-                if ($model->fields['raz_fusioninventory']) {
-                    PluginUninstallUninstall::deleteFusionInventoryLink(get_class($olditem), $olditem_id);
-                }
+            if ($plug->isActivated('fusioninventory') && $model->fields['raz_fusioninventory']) {
+                PluginUninstallUninstall::deleteFusionInventoryLink($olditem::class, $olditem_id);
             }
 
             // METHOD REPLACEMENT 1 : Purge
             switch ($model->fields['replace_method']) {
                 case self::METHOD_PURGE:
                     // Retrieve, Compute && Update NEW comment field
-                    if (!empty($newitem->fields['comment'])) {
-                        $comment = stripslashes($newitem->fields['comment']);
-                    } else {
-                        $comment = "";
-                    }
+                    $comment = empty($newitem->fields['comment']) ? "" : stripslashes((string) $newitem->fields['comment']);
 
                     $comment .= self::getCommentsForReplacement($olditem, true);
                     $comment .= "\n- " . __('See attached document', 'uninstall');
@@ -535,24 +518,18 @@ class PluginUninstallReplace extends CommonDBTM
                     if (isset($document_added) && $document_added) {
                         $olditem->delete(['id' => $olditem_id], true);
                     }
+
                     break;
 
                 case self::METHOD_DELETE_AND_COMMENT:
                 case self::METHOD_KEEP_AND_COMMENT:
                     // Retrieve && Compute comment for newitem (with olditem)
-                    if (!empty($newitem->fields['comment'])) {
-                        $commentnew = stripslashes($newitem->fields['comment']);
-                    } else {
-                        $commentnew = "";
-                    }
-                    $commentnew .= self::getCommentsForReplacement($olditem, true);
+                    $commentnew = empty($newitem->fields['comment']) ? "" : stripslashes((string) $newitem->fields['comment']);
 
+                    $commentnew .= self::getCommentsForReplacement($olditem, true);
                     // Retrieve && Compute comment for olditem (with newitem)
-                    if (!empty($olditem->getField('comment'))) {
-                        $commentold = stripslashes($olditem->getField('comment'));
-                    } else {
-                        $commentold = "";
-                    }
+                    $commentold = empty($olditem->getField('comment')) ? "" : stripslashes((string) $olditem->getField('comment'));
+
                     $commentold .= self::getCommentsForReplacement($newitem, false);
 
                     // Update comment for newitem
@@ -582,6 +559,7 @@ class PluginUninstallReplace extends CommonDBTM
                     if ($model->fields['replace_method'] == self::METHOD_DELETE_AND_COMMENT) {
                         $olditem->delete(['id' => $olditem_id], false, false);
                     }
+
                     break;
             }
 
@@ -607,6 +585,7 @@ class PluginUninstallReplace extends CommonDBTM
             foreach ($tab_ids as $olditem_id => $newitem_id) {
                 $uninstallArray[$olditem_id] = $olditem_id;
             }
+
             PluginUninstallUninstall::uninstall(
                 $type,
                 $model->getID(),
@@ -658,6 +637,7 @@ class PluginUninstallReplace extends CommonDBTM
                 $item->getField('otherserial'),
             );
         }
+
         return $string;
     }
 
@@ -726,6 +706,7 @@ class PluginUninstallReplace extends CommonDBTM
                 echo "<span class='green b'>" . $methods[self::METHOD_KEEP_AND_COMMENT] . "</span>";
                 break;
         }
+
         echo "</td></tr>";
 
         echo "<tr><td colspan='6'></td></tr>";
@@ -757,6 +738,7 @@ class PluginUninstallReplace extends CommonDBTM
             echo Dropdown::getDropdownName('glpi_states', $model->fields['states_id']);
             echo "</span>";
         }
+
         echo "</td>";
         echo "<td></td></tr>";
 
@@ -827,7 +809,7 @@ class PluginUninstallReplace extends CommonDBTM
         echo "<table class='tab_cadre_fixe' cellpadding='5'>";
 
         echo "<tr class='tab_bg_1 center'>";
-        count($tab_ids[$type]) > 1 ? $colspan = 5 : $colspan = 4;
+        $colspan = count($tab_ids[$type]) > 1 ? 5 : 4;
         echo "<th colspan='" . $colspan . "'>" . __('Choices for item to replace', 'uninstall') . "</th></tr>";
 
         echo "<tr class='tab_bg_1 center'>";
@@ -867,7 +849,7 @@ class PluginUninstallReplace extends CommonDBTM
 
                 echo "<td>";
                 $type::dropdown([
-                    'name'        => "newItems[$id]",
+                    'name'        => sprintf('newItems[%s]', $id),
                     'displaywith' => ['serial', 'otherserial'],
                     'url'         => $CFG_GLPI['root_doc'] . "/plugins/uninstall/ajax/dropdownReplaceFindDevice.php",
                     'used'        => array_keys($tab_ids[$type]),
@@ -973,6 +955,7 @@ class PluginUninstallReplace extends CommonDBTM
         foreach ($it as $data) {
             $docs[] = $data;
         }
+
         return $docs;
     }
 
@@ -1069,6 +1052,7 @@ class PluginUninstallReplace extends CommonDBTM
         foreach ($it as $data) {
             $tickets[] = $data;
         }
+
         return $tickets;
     }
 
@@ -1103,6 +1087,7 @@ class PluginUninstallReplace extends CommonDBTM
         foreach ($it as $data) {
             $netports[] = $data;
         }
+
         return $netports;
     }
 
@@ -1126,6 +1111,7 @@ class PluginUninstallReplace extends CommonDBTM
             if (!class_exists($itemtype)  || !is_a($itemtype, CommonDBTM::class, true)) {
                 continue;
             }
+
             $item = new $itemtype();
             if ($item->canView()) {
                 $datas = getAllDataFromTable(
@@ -1158,7 +1144,7 @@ class PluginUninstallReplace extends CommonDBTM
                 'itemtype' => $item->getType(),
             ],
         ]);
-        if (!count($iterator)) {
+        if (count($iterator) === 0) {
             //Get all item's tabs
             $tab = array_keys($item->defineTabs());
 
@@ -1171,6 +1157,7 @@ class PluginUninstallReplace extends CommonDBTM
             foreach ($iterator as $data) {
                 $tabs[] = $data['tabref'];
             }
+
             return $tabs;
         }
     }
